requirements.txt
scikit-learn
clearml
numpy
pandas
joblib
apache-airflow
dvc
uvicorn
requests
pydantic
fastapi
streamlit
httpx
pytest
